Test texture loader class
Created some duplicates and png to test the load class. I works ok but I
think i need to create a custom event for the loading progress because
the Game class need to receive the loading event also. At this point of
time, only the loading screen class can listen to the event.